+++ server/driver.go
@@ -21,7 +21,7 @@ import (
 // IDriver opens IContext.
 type IDriver interface {
 	// OpenCtx opens an IContext with connection id, client capability, collation, dbname and optionally the tls state.
-	OpenCtx(connID uint64, capability uint32, collation uint8, dbname string, tlsState *tls.ConnectionState) (*TiDBContext, error)
+	OpenCtx(connID uint64, capability uint32, collation uint16, dbname string, tlsState *tls.ConnectionState) (*TiDBContext, error)
 }
 
 //// PreparedStatement is the interface to use a prepared statement.
